Added Option make functions to OptionsUiModel.
