+++ OpenTESArena/src/Interface/OptionsPanel.cpp
@@ -107,377 +107,38 @@ OptionsPanel::OptionsPanel(Game &game)
 	this->tabButton = Button<OptionsPanel&, OptionsUiModel::Tab*, OptionsUiModel::Tab>(
 		OptionsUiController::onTabButtonSelected);
 
-	const auto &options = game.getOptions();
-
 	// Create graphics options.
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::WINDOW_MODE_NAME,
-		options.getGraphics_WindowMode(),
-		1,
-		Options::MIN_WINDOW_MODE,
-		Options::MAX_WINDOW_MODE,
-		std::vector<std::string> { "Window", "Borderless Full" },
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		auto &renderer = game.getRenderer();
-		options.setGraphics_WindowMode(value);
-
-		const Renderer::WindowMode mode = [value]()
-		{
-			switch (value)
-			{
-			case 0:
-				return Renderer::WindowMode::Window;
-			case 1:
-				return Renderer::WindowMode::BorderlessFull;
-			default:
-				DebugUnhandledReturnMsg(Renderer::WindowMode, std::to_string(value));
-			}
-		}();
-
-		renderer.setWindowMode(mode);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::FPS_LIMIT_NAME,
-		options.getGraphics_TargetFPS(),
-		5,
-		Options::MIN_FPS,
-		std::numeric_limits<int>::max(),
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setGraphics_TargetFPS(value);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::RESOLUTION_SCALE_NAME,
-		"Percent of the window resolution to use for software rendering.\nThis has a significant impact on performance.",
-		options.getGraphics_ResolutionScale(),
-		0.050,
-		Options::MIN_RESOLUTION_SCALE,
-		Options::MAX_RESOLUTION_SCALE,
-		2,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setGraphics_ResolutionScale(value);
-
-		// Resize the game world rendering.
-		auto &renderer = game.getRenderer();
-		const Int2 windowDimensions = renderer.getWindowDimensions();
-		const bool fullGameWindow = options.getGraphics_ModernInterface();
-		renderer.resize(windowDimensions.x, windowDimensions.y,
-			value, fullGameWindow);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::VERTICAL_FOV_NAME,
-		"Recommended 60.0 for classic mode.",
-		options.getGraphics_VerticalFOV(),
-		5.0,
-		Options::MIN_VERTICAL_FOV,
-		Options::MAX_VERTICAL_FOV,
-		1,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setGraphics_VerticalFOV(value);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::LETTERBOX_MODE_NAME,
-		"Determines the aspect ratio of the game UI. The weapon animation\nin modern mode is unaffected by this.",
-		options.getGraphics_LetterboxMode(),
-		1,
-		Options::MIN_LETTERBOX_MODE,
-		Options::MAX_LETTERBOX_MODE,
-		std::vector<std::string> { "16:10", "4:3", "Stretch" },
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		auto &renderer = game.getRenderer();
-		options.setGraphics_LetterboxMode(value);
-		renderer.setLetterboxMode(value);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::CURSOR_SCALE_NAME,
-		options.getGraphics_CursorScale(),
-		0.10,
-		Options::MIN_CURSOR_SCALE,
-		Options::MAX_CURSOR_SCALE,
-		1,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setGraphics_CursorScale(value);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::MODERN_INTERFACE_NAME,
-		"Modern mode uses a minimal interface with free-look.",
-		options.getGraphics_ModernInterface(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setGraphics_ModernInterface(value);
-
-		// If classic mode, make sure the player is looking straight forward.
-		// This is a restriction on the camera to retain the original feel.
-		const bool isModernMode = value;
-		if (!isModernMode)
-		{
-			auto &player = game.getGameState().getPlayer();
-			player.setDirectionToHorizon();
-		}
-
-		// Resize the game world rendering.
-		auto &renderer = game.getRenderer();
-		const Int2 windowDimensions = renderer.getWindowDimensions();
-		const bool fullGameWindow = isModernMode;
-		renderer.resize(windowDimensions.x, windowDimensions.y,
-			options.getGraphics_ResolutionScale(), fullGameWindow);
-	}));
-
-	this->graphicsOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::RENDER_THREADS_MODE_NAME,
-		"Determines the number of CPU threads to use for rendering.\nThis has a significant impact on performance.\nVery Low: one, Low: 1/4, Medium: 1/2, High: 3/4,\nVery High: all but one, Max: all",
-		options.getGraphics_RenderThreadsMode(),
-		1,
-		Options::MIN_RENDER_THREADS_MODE,
-		Options::MAX_RENDER_THREADS_MODE,
-		std::vector<std::string> { "Very Low", "Low", "Medium", "High", "Very High", "Max" },
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		auto &renderer = game.getRenderer();
-		options.setGraphics_RenderThreadsMode(value);
-		renderer.setRenderThreadsMode(value);
-	}));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeWindowModeOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeFpsLimitOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeResolutionScaleOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeVerticalFovOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeLetterboxModeOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeCursorScaleOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeModernInterfaceOption(game));
+	this->graphicsOptions.emplace_back(OptionsUiModel::makeRenderThreadsModeOption(game));
 
 	// Create audio options.
-	this->audioOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::SOUND_CHANNELS_NAME,
-		"Determines max number of concurrent sounds (including music).\nChanges are applied on next program start.",
-		options.getAudio_SoundChannels(),
-		1,
-		Options::MIN_SOUND_CHANNELS,
-		std::numeric_limits<int>::max(),
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setAudio_SoundChannels(value);
-	}));
-
-	this->audioOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::SOUND_RESAMPLING_NAME,
-		"Affects quality of sounds. Results may vary depending on OpenAL\nversion.",
-		options.getAudio_SoundResampling(),
-		1,
-		0,
-		Options::RESAMPLING_OPTION_COUNT - 1,
-		std::vector<std::string> { "Default", "Fastest", "Medium", "Best" },
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setAudio_SoundResampling(value);
-
-		// If the sound resampling extension is supported, update the audio manager sources.
-		auto &audioManager = game.getAudioManager();
-		if (audioManager.hasResamplerExtension())
-		{
-			audioManager.setResamplingOption(value);
-		}
-	}));
-
-	this->audioOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::IS_3D_AUDIO_NAME,
-		"Determines whether sounds in the game world have a 3D position.\nSet to false for classic behavior.",
-		options.getAudio_Is3DAudio(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setAudio_Is3DAudio(value);
-
-		auto &audioManager = game.getAudioManager();
-		audioManager.set3D(value);
-	}));
+	this->audioOptions.emplace_back(OptionsUiModel::makeSoundChannelsOption(game));
+	this->audioOptions.emplace_back(OptionsUiModel::makeSoundResamplingOption(game));
+	this->audioOptions.emplace_back(OptionsUiModel::makeIs3dAudioOption(game));
 
 	// Create input options.
-	this->inputOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::HORIZONTAL_SENSITIVITY_NAME,
-		options.getInput_HorizontalSensitivity(),
-		0.10,
-		Options::MIN_HORIZONTAL_SENSITIVITY,
-		Options::MAX_HORIZONTAL_SENSITIVITY,
-		1,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setInput_HorizontalSensitivity(value);
-	}));
-
-	this->inputOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::VERTICAL_SENSITIVITY_NAME,
-		"Only affects camera look in modern mode.",
-		options.getInput_VerticalSensitivity(),
-		0.10,
-		Options::MIN_VERTICAL_SENSITIVITY,
-		Options::MAX_VERTICAL_SENSITIVITY,
-		1,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setInput_VerticalSensitivity(value);
-	}));
-
-	this->inputOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::CAMERA_PITCH_LIMIT_NAME,
-		"Determines how far above or below the horizon the camera can\nlook in modern mode.",
-		options.getInput_CameraPitchLimit(),
-		5.0,
-		Options::MIN_CAMERA_PITCH_LIMIT,
-		Options::MAX_CAMERA_PITCH_LIMIT,
-		1,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setInput_CameraPitchLimit(value);
-
-		// Reset player view to forward.
-		auto &player = game.getGameState().getPlayer();
-		player.setDirectionToHorizon();
-	}));
-
-	this->inputOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::PIXEL_PERFECT_SELECTION_NAME,
-		"Changes entity selection so only clicks on opaque places are\nregistered, if enabled.",
-		options.getInput_PixelPerfectSelection(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setInput_PixelPerfectSelection(value);
-	}));
+	this->inputOptions.emplace_back(OptionsUiModel::makeHorizontalSensitivityOption(game));
+	this->inputOptions.emplace_back(OptionsUiModel::makeVerticalSensitivityOption(game));
+	this->inputOptions.emplace_back(OptionsUiModel::makeCameraPitchLimitOption(game));
+	this->inputOptions.emplace_back(OptionsUiModel::makePixelPerfectSelectionOption(game));
 
 	// Create miscellaneous options.
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::SHOW_COMPASS_NAME,
-		options.getMisc_ShowCompass(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_ShowCompass(value);
-	}));
-
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::SHOW_INTRO_NAME,
-		"Shows startup logo and related screens.",
-		options.getMisc_ShowIntro(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_ShowIntro(value);
-	}));
-
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::DoubleOption>(
-		OptionsUiModel::TIME_SCALE_NAME,
-		"Affects speed of gameplay. Lower this to simulate the speed of\nlower cycles in DOSBox.",
-		options.getMisc_TimeScale(),
-		0.050,
-		Options::MIN_TIME_SCALE,
-		Options::MAX_TIME_SCALE,
-		2,
-		[this](double value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_TimeScale(value);
-	}));
-
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::CHUNK_DISTANCE_NAME,
-		"Affects how many chunks away from the player chunks are\nsimulated and rendered.",
-		options.getMisc_ChunkDistance(),
-		1,
-		Options::MIN_CHUNK_DISTANCE,
-		std::numeric_limits<int>::max(),
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_ChunkDistance(value);
-	}));
-
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::STAR_DENSITY_NAME,
-		"Determines number of stars in the sky. Changes take effect the next\ntime stars are generated.",
-		options.getMisc_StarDensity(),
-		1,
-		Options::MIN_STAR_DENSITY_MODE,
-		Options::MAX_STAR_DENSITY_MODE,
-		std::vector<std::string> { "Classic", "Moderate", "High" },
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_StarDensity(value);
-	}));
-
-	this->miscOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::PLAYER_HAS_LIGHT_NAME,
-		"Whether the player has a light attached like in the original game.",
-		options.getMisc_PlayerHasLight(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_PlayerHasLight(value);
-	}));
+	this->miscOptions.emplace_back(OptionsUiModel::makeShowCompassOption(game));
+	this->miscOptions.emplace_back(OptionsUiModel::makeShowIntroOption(game));
+	this->miscOptions.emplace_back(OptionsUiModel::makeTimeScaleOption(game));
+	this->miscOptions.emplace_back(OptionsUiModel::makeChunkDistanceOption(game));
+	this->miscOptions.emplace_back(OptionsUiModel::makeStarDensityOption(game));
+	this->miscOptions.emplace_back(OptionsUiModel::makePlayerHasLightOption(game));
 
 	// Create developer options.
-	this->devOptions.emplace_back(std::make_unique<OptionsUiModel::BoolOption>(
-		OptionsUiModel::COLLISION_NAME,
-		"Enables player collision (not fully implemented yet).",
-		options.getMisc_Collision(),
-		[this](bool value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_Collision(value);
-	}));
-
-	this->devOptions.emplace_back(std::make_unique<OptionsUiModel::IntOption>(
-		OptionsUiModel::PROFILER_LEVEL_NAME,
-		"Displays varying levels of profiler information in the game world.",
-		options.getMisc_ProfilerLevel(),
-		1,
-		Options::MIN_PROFILER_LEVEL,
-		Options::MAX_PROFILER_LEVEL,
-		[this](int value)
-	{
-		auto &game = this->getGame();
-		auto &options = game.getOptions();
-		options.setMisc_ProfilerLevel(value);
-	}));
+	this->devOptions.emplace_back(OptionsUiModel::makeCollisionOption(game));
+	this->devOptions.emplace_back(OptionsUiModel::makeProfilerLevelOption(game));
 
 	// Set initial tab.
 	this->tab = OptionsUiModel::Tab::Graphics;
